refactor: 외부에서 사용하지 않는 상수의 접근자를 private로 변경
