+++ src/main/java/kitchenpos/menugroup/application/MenuGroupDao.java
@@ -1,7 +1,7 @@
 package kitchenpos.menugroup.application;
 
 import java.util.List;
-import kitchenpos.menugroup.application.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
