+++ src/test/java/kitchenpos/application/KitchenposFixture.java
@@ -3,12 +3,12 @@ package kitchenpos.application;
 import java.lang.reflect.Field;
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.menugroup.application.dto.MenuGroupResponse;
+import kitchenpos.menugroup.dto.MenuGroupResponse;
 import kitchenpos.product.application.dto.ProductResponse;
 import kitchenpos.ordertable.application.TableService;
 import kitchenpos.ordertable.application.dto.TableResponse;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menugroup.application.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuProducts;
 import kitchenpos.order.domain.OrderLineItem;
